# yembis_backend/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "RoadWatch",
    "version": "v4"
  },
  "paths": {
    "/api/camera": {
      "get": {
        "tags": [
          "Cameras"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Camera"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Cameras"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Camera"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Camera"
            }
          }
        }
      }
    },
    "/api/camera/{id}": {
      "get": {
        "tags": [
          "Cameras"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Camera"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Cameras"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Camera"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Cameras"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/category": {
      "get": {
        "tags": [
          "Categories"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Categorie"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Categories"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Categorie"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Categorie"
            }
          }
        }
      }
    },
    "/api/category/{id}": {
      "get": {
        "tags": [
          "Categories"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Categorie"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Categories"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Categorie"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Categories"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/locations": {
      "get": {
        "tags": [
          "Locations"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Location"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Locations"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Location"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/locations/{id}": {
      "get": {
        "tags": [
          "Locations"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Location"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Locations"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Location"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Locations"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/meting": {
      "get": {
        "tags": [
          "Metings"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Meting"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Metings"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Meting"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Meting"
            }
          }
        }
      }
    },
    "/api/meting/{id}": {
      "get": {
        "tags": [
          "Metings"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Meting"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Metings"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Meting"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Metings"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/speedlimit": {
      "get": {
        "tags": [
          "SpeedLimits"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/SpeedLimit"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "SpeedLimits"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SpeedLimit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/SpeedLimit"
            }
          }
        }
      }
    },
    "/api/speedlimit/{id}": {
      "get": {
        "tags": [
          "SpeedLimits"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/SpeedLimit"
            }
          }
        }
      },
      "put": {
        "tags": [
          "SpeedLimits"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SpeedLimit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "SpeedLimits"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "Camera": {
      "type": "object",
      "properties": {
        "cameraID": {
          "format": "int32",
          "type": "integer"
        },
        "macAddress": {
          "type": "string"
        }
      }
    },
    "Categorie": {
      "type": "object",
      "properties": {
        "categorieID": {
          "format": "int32",
          "type": "integer"
        },
        "typeVoertuig": {
          "type": "string"
        }
      }
    },
    "Location": {
      "type": "object",
      "properties": {
        "locationID": {
          "format": "int32",
          "type": "integer"
        },
        "cameraID": {
          "format": "int32",
          "type": "integer"
        },
        "gemeente": {
          "type": "string"
        },
        "straat": {
          "type": "string"
        },
        "longitude": {
          "type": "string"
        },
        "latitude": {
          "type": "string"
        },
        "vrachtwagenvrijzone": {
          "type": "boolean"
        },
        "beginPeriode": {
          "format": "date-time",
          "type": "string"
        },
        "eindPeriode": {
          "format": "date-time",
          "type": "string"
        },
        "current": {
          "type": "boolean"
        },
        "camera": {
          "$ref": "#/definitions/Camera"
        }
      }
    },
    "Meting": {
      "required": [
        "datumTijd",
        "snelheid"
      ],
      "type": "object",
      "properties": {
        "metingID": {
          "format": "int32",
          "type": "integer"
        },
        "cameraID": {
          "format": "int32",
          "type": "integer"
        },
        "categorieID": {
          "format": "int32",
          "type": "integer"
        },
        "locationID": {
          "format": "int32",
          "type": "integer"
        },
        "datumTijd": {
          "format": "date-time",
          "type": "string"
        },
        "snelheid": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "SpeedLimit": {
      "type": "object",
      "properties": {
        "limitID": {
          "format": "int32",
          "type": "integer"
        },
        "categorieID": {
          "format": "int32",
          "type": "integer"
        },
        "locationID": {
          "format": "int32",
          "type": "integer"
        },
        "snelheidslimiet": {
          "format": "int32",
          "type": "integer"
        },
        "categorie": {
          "$ref": "#/definitions/Categorie"
        },
        "location": {
          "$ref": "#/definitions/Location"
        }
      }
    }
  }
}